[EN-5197] [Multiplatform API (GraalVM)] Add subscription entity (#11)
* [EN-5194] [Multiplatform API (GraalVM)] Development of dxfeed-graal-native-api

+++ src/main/java/com/dxfeed/api/subscription/SubscriptionNative.java
@@ -6,15 +6,25 @@ import com.dxfeed.api.BaseNative;
 import com.dxfeed.api.DXFeed;
 import com.dxfeed.api.DXFeedEventListener;
 import com.dxfeed.api.DXFeedSubscription;
-import com.dxfeed.api.events.DxfgEventPointer;
 import com.dxfeed.api.events.DxfgCCharPointerPointer;
+import com.dxfeed.api.events.DxfgEventPointer;
 import com.dxfeed.api.exception.ExceptionHandlerReturnMinusOne;
 import com.dxfeed.api.osub.ObservableSubscriptionChangeListener;
 import com.dxfeed.event.EventType;
+import com.dxfeed.event.market.CandleExchangeMapper;
+import com.dxfeed.event.market.CandleMapper;
+import com.dxfeed.event.market.CandlePeriodMapper;
+import com.dxfeed.event.market.CandlePriceLevelMapper;
+import com.dxfeed.event.market.CandleSymbolMapper;
 import com.dxfeed.event.market.ListEventMapper;
+import com.dxfeed.event.market.QuoteMapper;
+import com.dxfeed.event.market.StringMapper;
+import com.dxfeed.event.market.TimeAndSaleMapper;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import java.util.concurrent.Executor;
 import org.graalvm.nativeimage.CurrentIsolate;
@@ -22,34 +32,28 @@ import org.graalvm.nativeimage.IsolateThread;
 import org.graalvm.nativeimage.c.CContext;
 import org.graalvm.nativeimage.c.function.CEntryPoint;
 import org.graalvm.nativeimage.c.type.CCharPointer;
+import org.graalvm.nativeimage.c.type.VoidPointer;
 
 @CContext(Directives.class)
 public class SubscriptionNative extends BaseNative {
 
-  private static final ListEventMapper EVENT_MAPPER = new ListEventMapper();
+  private static final ListEventMapper EVENT_MAPPER;
+  private static final Map<Long, DXFeedEventListener<EventType<?>>> EVENT_LISTENERS = new HashMap<>();
 
-  public DXFeedSubscription<?> constructorDXFeedSubscription(
-      final Class<?> eventType
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public DXFeedSubscription<?> constructorDXFeedSubscription(
-      final Class<?>... eventTypes
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public void attach(final DXFeed feed) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public void detach(final DXFeed feed) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public boolean isClosed() {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
+  static {
+    final StringMapper stringMapper = new StringMapper();
+    EVENT_MAPPER = new ListEventMapper(
+        new QuoteMapper(stringMapper),
+        new TimeAndSaleMapper(stringMapper),
+        new CandleMapper(
+            new CandleSymbolMapper(
+                stringMapper,
+                new CandlePeriodMapper(stringMapper),
+                new CandleExchangeMapper(),
+                new CandlePriceLevelMapper()
+            )
+        )
+    );
   }
 
   @CEntryPoint(
@@ -64,16 +68,6 @@ public class SubscriptionNative extends BaseNative {
     return EXECUTE_SUCCESSFULLY;
   }
 
-  public Set<Class<?>> getEventTypes() {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public boolean containsEventType(
-      final Class<?> eventType
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
   @CEntryPoint(
       name = "dxfg_subscription_clear",
       exceptionHandler = ExceptionHandlerReturnMinusOne.class
@@ -84,26 +78,6 @@ public class SubscriptionNative extends BaseNative {
   ) {
     getDxFeedSubscription(dxfgSubscription.getJavaObjectHandler()).clear();
     return EXECUTE_SUCCESSFULLY;
-  }
-
-  public Set<?> getSymbols() {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public Set<?> getDecoratedSymbols() {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public void setSymbols(
-      final Collection<?> symbols
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public void setSymbols(
-      final Object... symbols
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
   }
 
   @CEntryPoint(
@@ -172,16 +146,6 @@ public class SubscriptionNative extends BaseNative {
     return EXECUTE_SUCCESSFULLY;
   }
 
-  public Executor getExecutor() {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
-  public void setExecutor(
-      final Executor executor
-  ) {
-    throw new UnsupportedOperationException("It has not yet been implemented.");
-  }
-
   @CEntryPoint(
       name = "dxfg_subscription_add_event_listener",
       exceptionHandler = ExceptionHandlerReturnMinusOne.class
@@ -189,18 +153,20 @@ public class SubscriptionNative extends BaseNative {
   public static int addEventListener(
       final IsolateThread ignoreThread,
       final DxfgSubscription dxfgSubscription,
-      final DxfgEventListener dxfgEventListener
+      final DxfgEventListener dxfgEventListener,
+      final VoidPointer userData
   ) {
-    final DXFeedEventListener<EventType<?>> listener = events -> {
-      final DxfgEventPointer nativeEvents = EVENT_MAPPER.nativeObject(events);
-      final IsolateThread currentThread = CurrentIsolate.getCurrentThread();
-      final int size = events.size();
-      dxfgEventListener.getEventListener().invoke(currentThread, nativeEvents, size);
-      EVENT_MAPPER.delete(nativeEvents, size);
-    };
-    dxfgEventListener.setJavaObjectHandler(createJavaObjectHandler(listener));
-    getDxFeedSubscription(dxfgSubscription.getJavaObjectHandler())
-        .addEventListener(listener);
+    if (!EVENT_LISTENERS.containsKey(dxfgEventListener.rawValue())) {
+      final DXFeedEventListener<EventType<?>> listener = events -> {
+        final DxfgEventPointer nativeEvents = EVENT_MAPPER.nativeObject(events);
+        final IsolateThread currentThread = CurrentIsolate.getCurrentThread();
+        final int size = events.size();
+        dxfgEventListener.invoke(currentThread, nativeEvents, size, userData);
+        EVENT_MAPPER.delete(nativeEvents, size);
+      };
+      EVENT_LISTENERS.put(dxfgEventListener.rawValue(), listener);
+      getDxFeedSubscription(dxfgSubscription.getJavaObjectHandler()).addEventListener(listener);
+    }
     return EXECUTE_SUCCESSFULLY;
   }
 
@@ -215,12 +181,73 @@ public class SubscriptionNative extends BaseNative {
   ) {
     getDxFeedSubscription(dxfgSubscription.getJavaObjectHandler())
         .removeEventListener(
-            getDxFeedEventListener(
-                dxfgEventListener.getJavaObjectHandler()
-            )
+            EVENT_LISTENERS.remove(dxfgEventListener.rawValue())
         );
-    destroyJavaObjectHandler(dxfgEventListener.getJavaObjectHandler());
     return EXECUTE_SUCCESSFULLY;
+  }
+
+  public DXFeedSubscription<?> constructorDXFeedSubscription(
+      final Class<?> eventType
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public DXFeedSubscription<?> constructorDXFeedSubscription(
+      final Class<?>... eventTypes
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public void attach(final DXFeed feed) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public void detach(final DXFeed feed) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public boolean isClosed() {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public Set<Class<?>> getEventTypes() {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public boolean containsEventType(
+      final Class<?> eventType
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public Set<?> getSymbols() {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public void setSymbols(
+      final Collection<?> symbols
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public void setSymbols(
+      final Object... symbols
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public Set<?> getDecoratedSymbols() {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public Executor getExecutor() {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
+  }
+
+  public void setExecutor(
+      final Executor executor
+  ) {
+    throw new UnsupportedOperationException("It has not yet been implemented.");
   }
 
   public synchronized void addChangeListener(
